# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Score My Recipe",
    "description": "A tool to compute Green-Score of recipes.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/parse_text": {
      "post": {
        "summary": "Parse Text",
        "description": "Parse a text and return a list of ingredients with quantities and eventual modifiers",
        "operationId": "parse_text_v1_parse_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeParseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/origins": {
      "get": {
        "summary": "Get Origins",
        "description": "Get the list of origins available in the database\n\nNote: as the list is not too big, we let clients handle suggestions to users",
        "operationId": "get_origins_v1_origins_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language for the request (2 or 5 letter code)",
              "title": "Lang"
            },
            "description": "Language for the request (2 or 5 letter code)"
          },
          {
            "name": "include_synonyms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, include the synonyms of each item in the response.",
              "default": false,
              "title": "Include Synonyms"
            },
            "description": "If true, include the synonyms of each item in the response."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OriginsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/labels": {
      "get": {
        "summary": "Get Labels",
        "description": "Get the list of labels relevant for green-score computation",
        "operationId": "get_labels_v1_labels_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language for the request (2 or 5 letter code)",
              "title": "Lang"
            },
            "description": "Language for the request (2 or 5 letter code)"
          },
          {
            "name": "include_synonyms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, include the synonyms of each item in the response.",
              "default": false,
              "title": "Include Synonyms"
            },
            "description": "If true, include the synonyms of each item in the response."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabelsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/countries": {
      "get": {
        "summary": "Get Countries",
        "description": "Get the list of countries relevant for green-score computation",
        "operationId": "get_countries_v1_countries_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language for the request (2 or 5 letter code)",
              "title": "Lang"
            },
            "description": "Language for the request (2 or 5 letter code)"
          },
          {
            "name": "include_synonyms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, include the synonyms of each item in the response.",
              "default": false,
              "title": "Include Synonyms"
            },
            "description": "If true, include the synonyms of each item in the response."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountriesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingredients": {
      "get": {
        "summary": "Get Ingredients",
        "description": "Get the list of ingredients relevant for green-score computation",
        "operationId": "get_ingredients_v1_ingredients_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language for the request (2 or 5 letter code)",
              "title": "Lang"
            },
            "description": "Language for the request (2 or 5 letter code)"
          },
          {
            "name": "include_synonyms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, include the synonyms of each item in the response.",
              "default": false,
              "title": "Include Synonyms"
            },
            "description": "If true, include the synonyms of each item in the response."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/units": {
      "get": {
        "summary": "Get Units",
        "description": "Get the list of units available in the Open Food Facts units taxonomy\n\nNote: as the list is not too big, we let clients handle suggestions to users",
        "operationId": "get_units_v1_units_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language for the request (2 or 5 letter code)",
              "title": "Lang"
            },
            "description": "Language for the request (2 or 5 letter code)"
          },
          {
            "name": "include_synonyms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, include the synonyms of each item in the response.",
              "default": false,
              "title": "Include Synonyms"
            },
            "description": "If true, include the synonyms of each item in the response."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/suggest-scored-ingredient": {
      "get": {
        "summary": "Suggest Scored Ingredient",
        "description": "Suggest scored ingredient alternatives for an ingredient with no Agribalyse match.\n\nGiven a taxonomy id (typically one returned in ``missing_ingredient_ids``),\nwalk down the ingredients taxonomy and return the descendants that resolve\nto an Agribalyse row, so the user can pick a more specific alternative.",
        "operationId": "suggest_scored_ingredient_v1_suggest_scored_ingredient_get",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Language for the request (2 or 5 letter code)",
              "title": "Lang"
            },
            "description": "Language for the request (2 or 5 letter code)"
          },
          {
            "name": "include_synonyms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, include the synonyms of each item in the response.",
              "default": false,
              "title": "Include Synonyms"
            },
            "description": "If true, include the synonyms of each item in the response."
          },
          {
            "name": "taxonomy_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Taxonomy id of the ingredient to find alternatives for",
              "title": "Taxonomy Id"
            },
            "description": "Taxonomy id of the ingredient to find alternatives for"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestScoredIngredientResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingredient-references": {
      "get": {
        "summary": "Ingredient References",
        "description": "Return proposed Agribalyse and linked CIQUAL correspondences for a taxonomy ingredient.",
        "operationId": "ingredient_references_v1_ingredient_references_get",
        "parameters": [
          {
            "name": "taxonomy_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Taxonomy Id"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientReferencesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/agribalyse/foods": {
      "get": {
        "summary": "Agribalyse Foods",
        "description": "Search the existing environmental food catalog.",
        "operationId": "agribalyse_foods_v1_agribalyse_foods_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 8,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodReferencesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/nutrition/foods": {
      "get": {
        "summary": "Ciqual Foods",
        "description": "Search actual food records in the bundled ANSES CIQUAL 2025 catalog.",
        "operationId": "ciqual_foods_v1_nutrition_foods_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 8,
              "title": "Limit"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodReferencesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/nutrition/products": {
      "get": {
        "summary": "Off Products",
        "description": "Search real OFF products through the backend, including cache and timeout handling.",
        "operationId": "off_products_v1_nutrition_products_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en",
              "title": "Lang"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30,
              "minimum": 1,
              "default": 8,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodReferencesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/green-score": {
      "post": {
        "summary": "Green Score",
        "description": "Compute the green-score of a recipe given as a list of ingredients.",
        "operationId": "green_score_v1_green_score_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GreenScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenScoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/prepared-weight": {
      "post": {
        "summary": "Prepared Weight",
        "description": "Estimate prepared mass using a documented profile for the selected food and method.",
        "operationId": "prepared_weight_v1_prepared_weight_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreparedWeightRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreparedWeightResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/nutrition/analyze": {
      "post": {
        "summary": "Analyze Nutrition",
        "description": "Analyze one served recipe component with independent algorithm-2023 availability.",
        "operationId": "analyze_nutrition_v1_nutrition_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NutritionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RecipeParseRequest": {
        "properties": {
          "lang": {
            "type": "string",
            "title": "Lang",
            "description": "Language for the request (2 or 5 letter code)"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "lang",
          "text"
        ],
        "title": "RecipeParseRequest",
        "description": "Request model for parse_text endpoint",
        "examples": [
          {
            "lang": "en",
            "text": "200g of apple, 1 cup of wheat flour"
          }
        ]
      },
      "RecipeParseResponse": {
        "properties": {
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/RecipeIngredient"
            },
            "type": "array",
            "title": "Ingredients"
          }
        },
        "type": "object",
        "required": [
          "ingredients"
        ],
        "title": "RecipeParseResponse",
        "description": "Response model for parse_text endpoint",
        "examples": [
          {
            "ingredients": [
              {
                "codified_ingredient": "apple",
                "is_in_taxonomy": true,
                "quantity_g": 150.0,
                "taxonomy_id": "en:apple"
              },
              {
                "codified_ingredient": "wheat flour",
                "is_in_taxonomy": true,
                "quantity_g": 200.0,
                "taxonomy_id": "en:wheat-flour"
              }
            ]
          }
        ]
      },
...
    }
  }
}